pyproject.toml
[tool.poetry]
name = "screencap"
version = "0.1.7"
description = ""
authors = ["gravegrow <gravegrow@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
rich = "^13.7.0"
numpy = "^1.26.2"
opencv-python = "^4.9.0.80"
opencv-stubs = "^0.0.8"
signals = { git = "https://github.com/gravegrow/signals.git" }
types-python-xlib = "^0.33.0.20240106"
numba = "^0.58.1"
python-xlib = "^0.33"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
